@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import{createRoot} from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
@@ -7,16 +7,15 @@ import { ChakraProvider } from "@chakra-ui/react";
 import ChatProvider from "./Context/ChatProvider";
 import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(
-  <ChakraProvider>
-    <BrowserRouter>
-      <ChatProvider>
-        <App />
-      </ChatProvider>
-    </BrowserRouter>
-  </ChakraProvider>,
-  document.getElementById("root")
-);
+const root = createRoot(document.getElementById("root"));
+root.render( 
+<ChakraProvider>
+  <BrowserRouter>
+    <ChatProvider>
+      <App />
+    </ChatProvider>
+  </BrowserRouter>
+</ChakraProvider>);
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
