@@ -35,11 +35,12 @@ const sendMessage = asyncHandler(async (req, res) => {
     chat: chatId,
   };
 
+  console.log(newMessage);
   try {
     var message = await Message.create(newMessage);
 
-    message = await message.populate("sender", "name pic").execPopulate();
-    message = await message.populate("chat").execPopulate();
+    message = await message.populate("sender", "name pic");
+    message = await message.populate("chat");
     message = await User.populate(message, {
       path: "chat.users",
       select: "name pic email",
